+++ src/AppStack.ts
@@ -2,10 +2,10 @@ import { Stack, StackProps } from 'aws-cdk-lib';
 import { Construct } from 'constructs';
 import { Configurable } from './Configuration';
 
-interface MainStackProps extends StackProps, Configurable { }
+interface AppStackProps extends StackProps, Configurable { }
 
-export class MainStack extends Stack {
-  constructor(scope: Construct, id: string, private readonly props: MainStackProps) {
+export class AppStack extends Stack {
+  constructor(scope: Construct, id: string, private readonly props: AppStackProps) {
     super(scope, id, props);
 
     // TODO add resources here
