+++ test/main.test.ts
@@ -1,11 +1,3 @@
-import { App } from 'aws-cdk-lib';
-import { Template } from 'aws-cdk-lib/assertions';
-import { MyStack } from '../src/main';
-
-test('Snapshot', () => {
-  const app = new App();
-  const stack = new MyStack(app, 'test');
-
-  const template = Template.fromStack(stack);
-  expect(template.toJSON()).toMatchSnapshot();
+test('test', () => {
+  console.log('No tests yet...');
 });
